pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hytools"
version = "0.1.0-alpha"
description = "Central package for Armenian language corpus contracts, extraction, and normalization"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Armenian Corpus Core Contributors"},
]
keywords = ["armenian", "corpus", "linguistics", "nlp", "extraction"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.28",
    "beautifulsoup4>=4.12",
    "lxml>=4.9",
    "pyyaml>=6.0",
    "pymongo>=4.7",
    "feedparser>=6.0",
]

[project.optional-dependencies]
augmentation = [
    "numpy>=1.24",
    "scipy>=1.10",
    "scikit-learn>=1.3",
]
browser = [
    "selenium>=4.10",
]
huggingface = [
    "datasets>=2.14",
]
mongodb = [
    "pymongo>=4.5",
]
rss = [
    "feedparser>=6.0",
]
all = [
    "hytools[browser,huggingface,mongodb,rss]",
]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.10",
    "mypy>=1.0",
    "mongomock>=4.0",
]

[project.urls]
Repository = "https://github.com/RVogel101/hytools"
Documentation = "https://github.com/RVogel101/hytools/wiki"

[tool.setuptools.packages.find]
where = ["."]
include = [
    "hytools*",
    "cleaning*",
    "ingestion*",
    "augmentation*",
    "integrations*",
    "linguistics*",
    "ocr*",
    "core_contracts*",
]
exclude = ["tests*", "docs*", "scripts*", "notebooks*", "migration_exports*"]

[tool.pyright]
exclude = [
    "**/__pycache__",
    "**/.git",
    "**/archive",
    "**/legacy",
    "**/docs/archive",
    "**/.mypy_cache",
    "**/.pytest_cache",
]
reportMissingImports = "warning"
reportMissingModuleSource = "warning"
